# docs/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cloacina-server API",
    "description": "REST API for the Cloacina workflow-orchestration server. WebSocket protocols (execution events, accumulator ingest, reactor commands, substrate delivery) are specified in the WebSocket Protocol reference of the documentation site.",
    "contact": {
      "name": "Dylan Storey",
      "email": "dylan.storey@gmail.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.7.0"
  },
  "paths": {
...
    "/v1/tenants/{tenant_id}/executions": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "GET /tenants/:tenant_id/executions — list workflow executions.",
        "description": "**CLOACI-T-0594 / API-02:** accepts `?status=Failed` and\n`?workflow_name=foo` and `?limit=N&offset=M`. Previously these\nquery params were silently discarded.",
        "operationId": "list_executions",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "workflow",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Executions page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantListResponse_ExecutionSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/v1/tenants/{tenant_id}/executions/{exec_id}": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "GET /tenants/:tenant_id/executions/:id — get execution details.",
        "operationId": "get_execution",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exec_id",
            "in": "path",
            "description": "Execution UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid execution ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Execution not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/v1/tenants/{tenant_id}/executions/{exec_id}/events": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "GET /tenants/:tenant_id/executions/:id/events — execution event log.",
        "operationId": "get_execution_events",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exec_id",
            "in": "path",
            "description": "Execution UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution event log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionEventsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid execution ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/v1/tenants/{tenant_id}/workflows/{name}/execute": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "POST /tenants/:tenant_id/workflows/:name/execute — execute a workflow.",
        "description": "CLOACI-T-0580: execution is routed through `TenantRunnerCache`, which\nreturns (or constructs) a `DefaultRunner` bound to the tenant's\n`Database`. The execution row + every event lands in the tenant's\nschema, never the admin schema.",
        "operationId": "execute_workflow",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Workflow name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Execution scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Execution failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access or role denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Standardized error response body. Every non-2xx response from\n`cloacina-server` carries this shape; the request correlation ID is in\nthe `x-request-id` response header, not the body.",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (e.g. `invalid_pagination`,\n`workflow_not_found`)."
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
      "ExecuteRequest": {
        "type": "object",
        "description": "Request body for `POST /tenants/{tenant_id}/workflows/{name}/execute`.",
        "properties": {
          "context": {
            "description": "Optional JSON context to pass to the workflow."
          }
        }
      },
      "ExecuteResponse": {
        "type": "object",
        "description": "`202 Accepted` body for a scheduled workflow execution.",
        "required": [
          "execution_id",
          "workflow_name",
          "tenant_id",
          "status"
        ],
        "properties": {
          "execution_id": {
            "type": "string",
            "description": "UUID of the scheduled execution."
          },
          "status": {
            "type": "string",
            "description": "Always `\"scheduled\"` at accept time."
          },
          "tenant_id": {
            "type": "string"
          },
          "workflow_name": {
            "type": "string"
          }
        }
      },
      "ExecutionDetail": {
        "type": "object",
        "description": "`GET /tenants/{tenant_id}/executions/{id}` response.",
        "required": [
          "tenant_id",
          "execution_id",
          "status"
        ],
        "properties": {
          "execution_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          }
        }
      },
      "ExecutionEvent": {
        "type": "object",
        "description": "One row in the execution event log.",
        "required": [
          "id",
          "event_type",
          "created_at",
          "sequence_num"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "description": "RFC 3339 timestamp."
          },
          "event_data": {
            "type": [
              "string",
              "null"
            ],
            "description": "JSON-encoded additional data for the event; `null` when absent."
          },
          "event_type": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "Event UUID."
          },
          "sequence_num": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ExecutionEventsResponse": {
        "type": "object",
        "description": "`GET /tenants/{tenant_id}/executions/{id}/events` response.",
        "required": [
          "tenant_id",
          "execution_id",
          "events"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExecutionEvent"
            }
          },
          "execution_id": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          }
        }
      },
...
      "TenantListResponse_ExecutionSummary": {
        "type": "object",
        "description": "List envelope variant that retains a top-level `tenant_id`, used by\ntenant-scoped list endpoints for backward compatibility with operator\ndashboards that key off it.",
        "required": [
          "tenant_id",
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One row in the executions list.",
              "required": [
                "id",
                "workflow_name",
                "status",
                "started_at"
              ],
              "properties": {
                "completed_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "RFC 3339 timestamp; `null` while still running."
                },
                "id": {
                  "type": "string",
                  "description": "Execution UUID."
                },
                "started_at": {
                  "type": "string",
                  "description": "RFC 3339 timestamp."
                },
                "status": {
                  "type": "string"
                },
                "workflow_name": {
                  "type": "string"
                }
              }
            }
          },
          "tenant_id": {
            "type": "string"
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cloacina API key, sent as `Authorization: Bearer <key>`. Tenant scope is carried by the key itself; tenant-scoped routes additionally name the tenant in the path."
      }
    }
  },
...
}